DAOS-12655 test: do not land, reproduction attempt
Try to coerce dmg server set-logmasks DEBUG which should be default
for most pool creates and destroys.

And add a hack to ensure DEBUG log_mask is used in the engines when
running pool destroy in tearDown, which is where this intermittent
bug is observed.

Experiment with reducing /mnt/daos partition size (4->3GB) in this
particular test that only needs 1-2GiB of tmpfs. This is to save
memory for running engines in the memory-constrained VM environment.
i.e., to investigate a suspicion that laggy application performance
might be due to memory pressure (possibly starting to swap?) while
using tmpfs and running the daos_engines.

Skip-func-test-leap15: false
Quick-Functional: true
Test-tag: test_rebuild_with_io
Test-repeat: 10

Required-githooks: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>